requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain_community
PyPDF2
faiss-cpu
langchain_google_genai
gunicorn